# Conflicts:
#	src/Eccube/Command/UpdateSchemaDoctrineCommand.php
#	src/Eccube/Controller/Admin/Product/ProductController.php
#	src/Eccube/Controller/Install/InstallController.php
#	src/Eccube/DependencyInjection/Facade/AnnotationReaderFacade.php
#	src/Eccube/Doctrine/EventSubscriber/TaxRuleEventSubscriber.php
#	src/Eccube/Doctrine/ORM/Mapping/Driver/TraitProxyAttributeDriver.php
#	src/Eccube/Doctrine/ORM/Query/Extract.php
#	src/Eccube/Entity/AbstractEntity.php
#	src/Eccube/Entity/AuthorityRole.php
#	src/Eccube/Entity/BaseInfo.php
#	src/Eccube/Entity/Block.php
#	src/Eccube/Entity/BlockPosition.php
#	src/Eccube/Entity/Calendar.php
#	src/Eccube/Entity/Cart.php
#	src/Eccube/Entity/CartItem.php
#	src/Eccube/Entity/Category.php
#	src/Eccube/Entity/ClassCategory.php
#	src/Eccube/Entity/ClassName.php
#	src/Eccube/Entity/Csv.php
#	src/Eccube/Entity/Customer.php
#	src/Eccube/Entity/CustomerAddress.php
#	src/Eccube/Entity/CustomerFavoriteProduct.php
#	src/Eccube/Entity/Delivery.php
#	src/Eccube/Entity/DeliveryDuration.php
#	src/Eccube/Entity/DeliveryFee.php
#	src/Eccube/Entity/DeliveryTime.php
#	src/Eccube/Entity/Layout.php
#	src/Eccube/Entity/LoginHistory.php
#	src/Eccube/Entity/MailHistory.php
#	src/Eccube/Entity/MailTemplate.php
#	src/Eccube/Entity/Master/AbstractMasterEntity.php
#	src/Eccube/Entity/Member.php
#	src/Eccube/Entity/News.php
#	src/Eccube/Entity/Order.php
#	src/Eccube/Entity/OrderItem.php
#	src/Eccube/Entity/OrderPdf.php
#	src/Eccube/Entity/Page.php
#	src/Eccube/Entity/PageLayout.php
#	src/Eccube/Entity/Payment.php
#	src/Eccube/Entity/PaymentOption.php
#	src/Eccube/Entity/Plugin.php
#	src/Eccube/Entity/PointRateTrait.php
#	src/Eccube/Entity/Product.php
#	src/Eccube/Entity/ProductCategory.php
#	src/Eccube/Entity/ProductClass.php
#	src/Eccube/Entity/ProductImage.php
#	src/Eccube/Entity/ProductStock.php
#	src/Eccube/Entity/ProductTag.php
#	src/Eccube/Entity/Shipping.php
#	src/Eccube/Entity/Tag.php
#	src/Eccube/Entity/TaxRule.php
#	src/Eccube/Entity/Template.php
#	src/Eccube/Form/Extension/DoctrineOrmExtension.php
#	src/Eccube/Form/Type/Admin/MasterdataType.php
#	src/Eccube/Kernel.php
#	src/Eccube/Service/Calculator/OrderItemCollection.php
#	src/Eccube/Service/CartService.php
#	src/Eccube/Service/PluginService.php
#	src/Eccube/Service/PurchaseFlow/ItemCollection.php
#	src/Eccube/Service/SchemaService.php

+++ src/Eccube/Service/PointHelper.php
@@ -18,6 +18,7 @@ use Eccube\Entity\ItemHolderInterface;
 use Eccube\Entity\Master\OrderItemType;
 use Eccube\Entity\Master\TaxDisplayType;
 use Eccube\Entity\Master\TaxType;
+use Eccube\Entity\Order;
 use Eccube\Entity\OrderItem;
 use Eccube\Repository\BaseInfoRepository;
 use Eccube\Service\PurchaseFlow\Processor\PointProcessor;
@@ -64,7 +65,7 @@ class PointHelper
     /**
      * ポイントを金額に変換する.
      *
-     * @param $point ポイント
+     * @param string $point ポイント
      *
      * @return string 金額
      *
@@ -75,13 +76,13 @@ class PointHelper
     {
         $BaseInfo = $this->baseInfoRepository->get();
 
-        return bcmul($point, $BaseInfo->getPointConversionRate(), 0);
+        return bcmul($point, (string) $BaseInfo->getPointConversionRate(), 0);
     }
 
     /**
      * ポイントを値引き額に変換する. マイナス値を返す.
      *
-     * @param $point ポイント
+     * @param string $point ポイント
      *
      * @return string 金額
      *
@@ -96,7 +97,7 @@ class PointHelper
     /**
      * 金額をポイントに変換する.
      *
-     * @param $price
+     * @param string $price
      *
      * @return string ポイント
      *
@@ -107,17 +108,26 @@ class PointHelper
     {
         $BaseInfo = $this->baseInfoRepository->get();
 
-        return bcfloor(bcdiv($price, $BaseInfo->getPointConversionRate(), 4));
+        return bcfloor(bcdiv($price, (string) $BaseInfo->getPointConversionRate(), 4));
     }
 
     /**
      * 明細追加処理.
      *
      * @param ItemHolderInterface $itemHolder
-     * @param int $discount
+     * @param string $discount
+     *
+     * @return void
+     *
+     * @throws \Exception
      */
     public function addPointDiscountItem(ItemHolderInterface $itemHolder, $discount)
     {
+        // 注文明細以外は処理しない.
+        if ($itemHolder instanceof Order === false) {
+            return;
+        }
+
         $DiscountType = $this->entityManager->find(OrderItemType::class, OrderItemType::POINT);
         $TaxInclude = $this->entityManager->find(TaxDisplayType::class, TaxDisplayType::INCLUDED);
         $Taxation = $this->entityManager->find(TaxType::class, TaxType::NON_TAXABLE);
@@ -136,9 +146,9 @@ class PointHelper
         $OrderItem->setProductName($DiscountType->getName())
             ->setPrice($discount)
             ->setPointRate($pointRate)
-            ->setQuantity(1)
-            ->setTax(0)
-            ->setTaxRate(0)
+            ->setQuantity('1')
+            ->setTax('0')
+            ->setTaxRate('0')
             ->setRoundingType(null)
             ->setOrderItemType($DiscountType)
             ->setTaxDisplayType($TaxInclude)
@@ -152,28 +162,44 @@ class PointHelper
      * 既存のポイント明細を削除する.
      *
      * @param ItemHolderInterface $itemHolder
+     *
+     * @return void
      */
     public function removePointDiscountItem(ItemHolderInterface $itemHolder)
     {
-        foreach ($itemHolder->getItems() as $item) {
-            if ($item->getProcessorName() == PointProcessor::class) {
-                $itemHolder->removeOrderItem($item);
-                $this->entityManager->remove($item);
+        if ($itemHolder instanceof Order) {
+            foreach ($itemHolder->getItems() as $item) {
+                if ($item instanceof OrderItem && $item->getProcessorName() == PointProcessor::class) {
+                    $itemHolder->removeOrderItem($item);
+                    $this->entityManager->remove($item);
+                }
             }
         }
     }
 
+    /**
+     * @param ItemHolderInterface $itemHolder
+     * @param string $point
+     *
+     * @return void
+     */
     public function prepare(ItemHolderInterface $itemHolder, $point)
     {
         // ユーザの保有ポイントを減算
         $Customer = $itemHolder->getCustomer();
-        $Customer->setPoint($Customer->getPoint() - $point);
+        $Customer->setPoint(bcsub($Customer->getPoint(), $point));
     }
 
+    /**
+     * @param ItemHolderInterface $itemHolder
+     * @param string $point
+     *
+     * @return void
+     */
     public function rollback(ItemHolderInterface $itemHolder, $point)
     {
         // 利用したポイントをユーザに戻す.
         $Customer = $itemHolder->getCustomer();
-        $Customer->setPoint($Customer->getPoint() + $point);
+        $Customer->setPoint(bcadd($Customer->getPoint(), $point));
     }
 }
